pyproject.toml
[tool.poetry]
name = "book-ims"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
sqlalchemy = "^2.0.31"
psycopg2 = "^2.9.9"
fastapi-sessions = "^0.3.2"
pyjwt = "^2.8.0"
pytest = "^8.2.2"
pandas = "^2.2.2"
sentence-transformers = "^3.0.1"
pinecone-client = "^5.0.0"
langchain = "^0.2.10"
uvicorn = "^0.30.3"
sse-starlette = "^2.1.2"
streamlit = "^1.36.0"
requests = "^2.32.3"
langchain-ollama = "^0.1.0"
sseclient = "^0.0.27"
langchain-community = "^0.2.9"
langchain-experimental = "^0.0.63"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
